// src/shared-components/case-study/color-typography/color-typography.d.ts
export interface IColorWithTypography {
  title: string;
  fontFamilyTitle: string;
  fontFamilyVariant: string;
  labelFontWeights: string;
  colorItemOne: {
    customStyles: string;
    colorNameOnCard: string;
  };
  colorItemTwo: {
    customStyles: string;
    colorNameOnCard: string;
  };
  colorItemThree: {
    customStyles: string;
    colorNameOnCard: string;
  };
  colorItemFour: {
    customStyles: string;
    colorNameOnCard: string;
  };
}